dinput: Always send both "Type Specific" and "Set Effect" reports on initial Download.
Fixes the Download method not sending an initial type specific
HID PID report in the rare cases where all type specific params
are set to 0 through the initial SetParameters call, so they aren't
considered as modified.
The "Set Envelope" report is an exception in this case, as the
effect is not used when all params are set to 0.

FH5 is affected by this issue, since it initially sets the
direction of its constant effect to 0, which translates to 0
magnitude in dinput.