winebus.sys: Add devtype operand to get_device_subsystem_info().
Narrow the search for a parent subsystem to a specific devtype.
For the usb subsystem in particular, only the "usb_device"
devtype includes all the necessary attributes, so no need to
iterate through the usb interfaces.

Additionally fixes an issue with the device revision (bcdDevice)
not being extracted from the usb subsystem, causing the bcdHID
to be extracted in error.

Also fixes an issue with the product name including the manufacturer
in the same string.
Both attributes are correctly extracted from the usb subsystem.