winebus.sys: When no serial number is available, use bus_num and port_path when generating instance_id.
If there are identical devices with the same serial numbers (unlikely
and goes against the specification), use SN only for the first device
with index 0. Otherwise use bus_num and port_path for the rest.

To make sure the instance_id is unique, use index if there is no bus_num
(and probably no port_path) available.