winebus.sys: Make container_id consistent through Wine launches.
container_id uniquely identifies a device on Windows, and some
apps store it for further reference. Changing this id on every
Wine run, may cause such app to decide that it is a new device
and not use its previously stored settings (button assignments,
FFB settings, etc.).
Additionally on Windows container_id identifies the entire device
incl. its interfaces and collections, i.e. it is the same for all
these instances (in USB and HID).

When generating container_id, use any of the following device
specific parameters, when available: serial number, or uid
(macos LocationID), or bus_num and port_path (Linux bus_udev
backend). If none of these are available use input and index,
which are interface specific, but will still generate an unique
id.