dinput: Make guidInstance consistent through Wine launches.
Many apps use this guid to uniquely identify a dinput device.
Making this guid solely dependent on the device enumeration
order (through the rawinput handle), may result in the same
guid being reused for another device on the next Wine launch.

To avoid this, when generating the device guidInstance use the
device_id and instance_id from DevicePath. These ids are consistent
through runs (depending on the parameters available to winebus),
and uniquely identify a HID device on the system.

Also change hid_joystick_guid.Data4 to a value similar to the one
used in Windows.